requirements.txt
langchain
requests
python-dotenv
google-generativeai
langchain_google_genai